requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.30.0
SQLAlchemy[asyncio]==2.0.36
imap-tools==1.6.0
anyio==4.4.0
cryptography==43.0.0
beautifulsoup4==4.12.3
pydantic==1.10.14
aiosqlite==0.20.0
email-validator==2.2.0
httpx==0.27.2
PyJWT==2.8.0
bcrypt>=4.0,<5
